ready for the solver

+++ bh_fdr.hpp
@@ -7,18 +7,18 @@ using namespace std;
 
 #define all(t) begin(t), end(t)
 #define sp << " " <<
-typedef pair<double,int> p_ind;
+typedef pair<float,int> p_ind;
 typedef vector<p_ind> sorted_p;
 
-vector<double> ecdf(unsigned int p_size) {
-  vector<double> v(p_size);
+vector<float> ecdf(unsigned int p_size) {
+  vector<float> v(p_size);
   iota(all(v),1);
-  for_each(all(v),[=](double& x){x /= p_size;});
+  for_each(all(v),[=](float& x){x /= p_size;});
   return v;
 }
-bool greater1(double x) {return x>1;}
+bool greater1(float x) {return x>1;}
 
-void bh_fdr(const vector<double>& p, vector<double>& q) {
+void bh_fdr(const vector<float>& p, vector<float>& q) {
   int T = p.size();
   sorted_p pvals(T);
   for (int i = 0; i < T;i++) {
@@ -26,15 +26,15 @@ void bh_fdr(const vector<double>& p, vector<double>& q) {
     pvals[i].second=i;
   }
   sort(all(pvals));
-  vector<double> ecdffactor;
+  vector<float> ecdffactor;
   ecdffactor = ecdf(T);
-  vector<double> pvals_corrected_raw(T);
+  vector<float> pvals_corrected_raw(T);
   transform(all(pvals), begin(ecdffactor), begin(pvals_corrected_raw),
-  	    [](p_ind pv, double x){return pv.first/x;});
+  	    [](p_ind pv, float x){return pv.first/x;});
   //pvals_corrected = np.minimum.accumulate(pvals_corrected_raw[::-1])[::-1]
   //first reverse pvals_corrected_raw
   reverse(all(pvals_corrected_raw));
-  vector<double> pvals_corrected(T);
+  vector<float> pvals_corrected(T);
   pvals_corrected[0]=pvals_corrected_raw[0];
   for (int i=1;i<T;i++) {
     pvals_corrected[i]=min(pvals_corrected_raw[i], pvals_corrected[i-1]);
@@ -43,7 +43,7 @@ void bh_fdr(const vector<double>& p, vector<double>& q) {
   reverse(all(pvals_corrected));
   //replace all values greater than 1 by 1
   replace_if(all(pvals_corrected),greater1,1);
-  vector<pair<int,double>> qvals(T);
+  vector<pair<int,float>> qvals(T);
   for (int i=0;i<T;i++) {
     qvals[i] = make_pair(pvals[i].second, pvals_corrected[i]);
   }
