+++ kstwo.hpp
@@ -7,12 +7,12 @@ using namespace std;
 #define EPS1 0.001
 #define EPS2 1.0e-8
 
-double probks(double alam) {
+float probks(float alam) {
 //Kolmogorov-Smirnov probability function.
-  double fac=2.0,sum=0.0,termbf=0.0;
-  double a2 = -2.0*alam*alam;
+  float fac=2.0,sum=0.0,termbf=0.0;
+  float a2 = -2.0*alam*alam;
   for (int j=1;j<=100;j++) {
-    double term=fac*exp(a2*j*j);
+    float term=fac*exp(a2*j*j);
     sum += term;
     if (fabs(term) <= EPS1*termbf || fabs(term) <= EPS2*sum) return sum;
     fac = -fac;//Alternating signs in sum.
@@ -22,7 +22,7 @@ double probks(double alam) {
   //Get here only by failing to converge.
 }
 
-pair<double, double> kstwo(const vector<double>& data1, const vector<double>& data2) {
+pair<float, float> kstwo(const vector<float>& data1, const vector<float>& data2) {
 /*
 Given an array data1[1..n1] , and an array data2[1..n2] , 
 this routine returns the K–
@@ -36,11 +36,11 @@ The arrays data1 and data2 are modified by being sorted into ascending order.
   //presorted to avoid repeated sorting
   //sort(all(data2));
   //presorted
-  double en1=data1.size();
-  double en2=data2.size();
-  double d=0.0;
+  float en1=data1.size();
+  float en2=data2.size();
+  float d=0.0;
   unsigned long j1=0,j2=0;
-  double d1,d2,dt,en,fn1=0.0,fn2=0.0;
+  float d1,d2,dt,en,fn1=0.0,fn2=0.0;
   //data1 en data2 are 0-indexed
   while (j1 < data1.size() && j2 < data2.size()) {
     if ((d1=data1[j1]) <= (d2=data2[j2])) fn1=j1++/en1;
@@ -48,6 +48,6 @@ The arrays data1 and data2 are modified by being sorted into ascending order.
     if ((dt=fabs(fn2-fn1)) > d) d=dt;//the max
   }
   en=sqrt(en1*en2/(en1+en2));
-  double prob=probks((en+0.12+0.11/en)*d);
+  float prob=probks((en+0.12+0.11/en)*d);
   return make_pair(d, prob);
 }
